Refactor logic for matchers to simplify main.go; Add matchers files based on categories of files.
