+++ main.go
@@ -8,11 +8,10 @@ package main
 import (
 	"archive/tar"
 	"archive/zip"
-	"bytes"
+	"encoding/json"
 	"flag"
 	"fmt"
 	"io"
-	"encoding/json"
 	"os"
 	"path/filepath"
 	"strings"
@@ -29,574 +28,6 @@ type fileMatcher struct {
 	minLen   int
 	match    func([]byte, int, int) bool
 	describe func([]byte, int, int, *os.File) string
-}
-
-var matchers = []fileMatcher{
-	{
-		name:   "elf",
-		minLen: 45,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb >= 45 && HasPrefix(b, "\x7FELF")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "Elf file " + doElf(b)
-		},
-	},
-	{
-		name:   "ar",
-		minLen: 8,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb >= 8 && HasPrefix(b, "!<arch>\n")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "ar archive"
-		},
-	},
-	{
-		name:   "png",
-		minLen: 29,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 28 && HasPrefix(b, "\x89PNG\x0d\x0a\x1a\x0a")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "PNG image data"
-		},
-	},
-	{
-		name:   "gif",
-		minLen: 17,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 16 && (HasPrefix(b, "GIF87a") || HasPrefix(b, "GIF89a"))
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "GIF image data"
-		},
-	},
-	{
-		name:   "jpeg",
-		minLen: 33,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 32 && HasPrefix(b, "\xff\xd8")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "JPEG / jpg image data"
-		},
-	},
-	{
-		name:   "java-class",
-		minLen: 9,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 8 && HasPrefix(b, "\xca\xfe\xba\xbe")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "Java class file"
-		},
-	},
-	{
-		name:   "dex",
-		minLen: 9,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 8 && HasPrefix(b, "dex\n")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "Android dex file"
-		},
-	},
-	{
-		name:   "tar",
-		minLen: 501,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 500 && Equal(b[257:262], "ustar")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return doTar(file)
-		},
-	},
-	{
-		name:   "zip",
-		minLen: 6,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 5 && HasPrefix(b, "PK\x03\x04")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return doZip(file)
-		},
-	},
-	{
-		name:   "vmdk",
-		minLen: 4,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb >= 4 && HasPrefix(b, "KDMV")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "VMware virtual disk"
-		},
-	},
-	{
-		name:   "bzip2",
-		minLen: 5,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 4 && HasPrefix(b, "BZh")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "bzip2 compressed data"
-		},
-	},
-	{
-		name:   "xz",
-		minLen: 6,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb >= 6 && HasPrefix(b, "\xFD7zXZ\x00")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "XZ compressed data"
-		},
-	},
-	{
-		name:   "zstd",
-		minLen: 4,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb >= 4 && HasPrefix(b, "\x28\xB5\x2F\xFD")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "Zstandard compressed data"
-		},
-	},
-	{
-		name:   "lz4",
-		minLen: 4,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb >= 4 && HasPrefix(b, "\x04\x22\x4D\x18")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "LZ4 compressed data"
-		},
-	},
-	{
-		name:   "lzip",
-		minLen: 4,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb >= 4 && HasPrefix(b, "LZIP")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "LZIP compressed data"
-		},
-	},
-	{
-		name:   "gzip",
-		minLen: 11,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 10 && HasPrefix(b, "\x1f\x8b")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "gzip compressed data"
-		},
-	},
-	{
-		name:   "wasm",
-		minLen: 4,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb >= 4 && HasPrefix(b, "\x00asm")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "WebAssembly binary"
-		},
-	},
-	{
-		name:   "macho",
-		minLen: 33,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 32 && Equal(b[1:4], "\xfa\xed\xfe")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "Mach-O"
-		},
-	},
-	{
-		name:   "flv",
-		minLen: 3,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb >= 3 && HasPrefix(b, "FLV")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "FLV video file"
-		},
-	},
-	{
-		name:   "matroska",
-		minLen: 4,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb >= 4 && HasPrefix(b, "\x1A\x45\xDF\xA3") && bytes.Contains(b, []byte("matroska"))
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "Matroska video file"
-		},
-	},
-	{
-		name:   "webm",
-		minLen: 4,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb >= 4 && HasPrefix(b, "\x1A\x45\xDF\xA3") && bytes.Contains(b, []byte("webm"))
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "WebM video file"
-		},
-	},
-	{
-		name:   "ogg",
-		minLen: 37,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 36 && HasPrefix(b, "OggS\x00\x02")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "Ogg data"
-		},
-	},
-	{
-		name:   "wav",
-		minLen: 33,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 32 && HasPrefix(b, "RIF") && Equal(b[8:16], "WAVEfmt ")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "WAV audio"
-		},
-	},
-	{
-		name:   "woff",
-		minLen: 4,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb >= 4 && HasPrefix(b, "wOFF")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "WOFF font"
-		},
-	},
-	{
-		name:   "woff2",
-		minLen: 4,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb >= 4 && HasPrefix(b, "wOF2")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "WOFF2 font"
-		},
-	},
-	{
-		name:   "ttf",
-		minLen: 13,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 12 && HasPrefix(b, "\x00\x01\x00\x00")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "TrueType font"
-		},
-	},
-	{
-		name:   "ttf-collection",
-		minLen: 13,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 12 && HasPrefix(b, "ttcf\x00")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "TrueType font collection"
-		},
-	},
-	{
-		name:   "llvm-bitcode",
-		minLen: 5,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 4 && HasPrefix(b, "BC\xc0\xde")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "LLVM IR bitcode"
-		},
-	},
-	{
-		name:   "parquet",
-		minLen: 4,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb >= 4 && HasPrefix(b, "PAR1")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "Parquet data"
-		},
-	},
-	{
-		name:   "avro",
-		minLen: 4,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb >= 4 && HasPrefix(b, "Obj\x01")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "Avro data"
-		},
-	},
-	{
-		name:   "pem",
-		minLen: len("-----BEGIN CERTIFICATE-----"),
-		match: func(b []byte, lenb int, magic int) bool {
-			return HasPrefix(b, "-----BEGIN CERTIFICATE-----")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "PEM certificate"
-		},
-	},
-	{
-		name:   "pe",
-		minLen: 64,
-		match: func(b []byte, lenb int, magic int) bool {
-			return magic != -1 && HasPrefix(b, "MZ") && magic < lenb-4 &&
-				Equal(b[magic:magic+4], "\x50\x45\x00\x00")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return describePE(b, magic)
-		},
-	},
-	{
-		name:   "iso9660",
-		minLen: 0x8006,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb >= 0x8006 && Equal(b[0x8001:0x8006], "CD001")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "ISO 9660 CD-ROM filesystem"
-		},
-	},
-	{
-		name:   "bmp",
-		minLen: 51,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 50 && HasPrefix(b, "BM") && Equal(b[6:10], "\x00\x00\x00\x00")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "BMP image"
-		},
-	},
-	{
-		name:   "pdf",
-		minLen: 51,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 50 && HasPrefix(b, "\x25\x50\x44\x46")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "PDF image"
-		},
-	},
-	{
-		name:   "tiff",
-		minLen: 17,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 16 &&
-				(HasPrefix(b, "\x49\x49\x2a\x00") || HasPrefix(b, "\x4D\x4D\x00\x2a"))
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "TIFF image data"
-		},
-	},
-	{
-		name:   "mp3",
-		minLen: 17,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 16 &&
-				(HasPrefix(b, "ID3") || HasPrefix(b, "\xff\xfb") || HasPrefix(b, "\xff\xf3") || HasPrefix(b, "\xff\xf2"))
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "MP3 audio file"
-		},
-	},
-	{
-		name:   "mp4",
-		minLen: 17,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 16 &&
-				(HasPrefix(b, "\x00\x00\x00\x20\x66\x74\x79\x70") || HasPrefix(b, "\x00\x00\x00\x18\x66\x74\x79\x70") || HasPrefix(b, "\x00\x00\x00\x14\x66\x74\x79\x70"))
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "MP4 video file"
-		},
-	},
-	{
-		name:   "rar",
-		minLen: 17,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 16 && HasPrefix(b, "\x52\x61\x72\x21\x1A\x07\x01\x00")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "RAR archive data"
-		},
-	},
-	{
-		name:   "7zip",
-		minLen: 17,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 16 && HasPrefix(b, "\x37\x7A\xBC\xAF\x27\x1C")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "7zip archive data"
-		},
-	},
-	{
-		name:   "ico",
-		minLen: 17,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 16 && HasPrefix(b, "\x00\x00\x01\x00")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "MS Windows icon resource"
-		},
-	},
-	{
-		name:   "sqlite",
-		minLen: 17,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 16 && HasPrefix(b, "\x53\x51\x4C\x69\x74\x65\x20\x66\x6F\x72\x6D\x61\x74\x20\x33\x00")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "SQLite database"
-		},
-	},
-	{
-		name:   "pcapng",
-		minLen: 17,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 16 && HasPrefix(b, "\x0A\x0D\x0D\x0A")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "PCAP-ng capture file"
-		},
-	},
-	{
-		name:   "pcap",
-		minLen: 17,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 16 &&
-				(HasPrefix(b, "\xD4\xC3\xB2\xA1") || HasPrefix(b, "\xA1\xB2\xC3\xD4") || HasPrefix(b, "\x4D\x3C\xB2\xA1") || HasPrefix(b, "\xA1\xB2\x3C\x4D"))
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "PCAP capture file"
-		},
-	},
-	{
-		name:   "flac",
-		minLen: 17,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 16 && HasPrefix(b, "\x66\x4C\x61\x43")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "FLAC audio format"
-		},
-	},
-	{
-		name:   "tdf",
-		minLen: 17,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 16 && HasPrefix(b, "\x54\x44\x46\x24")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "Telegram Desktop file"
-		},
-	},
-	{
-		name:   "tdef",
-		minLen: 17,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 16 && HasPrefix(b, "\x54\x44\x45\x46")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "Telegram Desktop encrypted file"
-		},
-	},
-	{
-		name:   "cab",
-		minLen: 17,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 16 && HasPrefix(b, "\x4D\x53\x43\x46")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "Microsoft Cabinet file"
-		},
-	},
-	{
-		name:   "psd",
-		minLen: 17,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 16 && HasPrefix(b, "\x38\x42\x50\x53")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "Photoshop document"
-		},
-	},
-	{
-		name:   "avi",
-		minLen: 33,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 32 && HasPrefix(b, "RIF") && Equal(b[8:11], "AVI")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "AVI file"
-		},
-	},
-	{
-		name:   "ole",
-		minLen: 33,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 32 && HasPrefix(b, "\xD0\xCF\x11\xE0\xA1\xB1\x1A\xE1")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "Microsoft Office (Legacy format)"
-		},
-	},
-	{
-		name:   "webp",
-		minLen: 33,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 32 && HasPrefix(b, "RIF") && Equal(b[8:12], "WEBP")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "Google Webp file"
-		},
-	},
-	{
-		name:   "rtf",
-		minLen: 33,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 32 && HasPrefix(b, "\x7B\x5C\x72\x74\x66\x31")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "Rich Text Format"
-		},
-	},
-	{
-		name:   "html",
-		minLen: 33,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 32 && (HasPrefix(b, "<!DOCTYPE html") || HasPrefix(b, "<head>"))
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "HTML document"
-		},
-	},
-	{
-		name:   "xml",
-		minLen: 33,
-		match: func(b []byte, lenb int, magic int) bool {
-			return lenb > 32 && HasPrefix(b, "<?xml version")
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return "XML document"
-		},
-	},
-	{
-		name:   "text",
-		minLen: 1,
-		match: func(b []byte, lenb int, magic int) bool {
-			return isText(b)
-		},
-		describe: func(b []byte, lenb int, magic int, file *os.File) string {
-			return describeText(b)
-		},
-	},
 }
 
 func main() {
