+++ matchers.go
@@ -1,0 +1,60 @@
+package main
+
+var matchers = []fileMatcher{
+	matcherElf,
+	matcherAr,
+	matcherPng,
+	matcherGif,
+	matcherJpeg,
+	matcherJavaClass,
+	matcherDex,
+	matcherTar,
+	matcherZip,
+	matcherVmdk,
+	matcherBzip2,
+	matcherXz,
+	matcherZstd,
+	matcherLz4,
+	matcherLzip,
+	matcherGzip,
+	matcherWasm,
+	matcherMacho,
+	matcherFlv,
+	matcherMatroska,
+	matcherWebm,
+	matcherOgg,
+	matcherWav,
+	matcherWoff,
+	matcherWoff2,
+	matcherTtf,
+	matcherTtfCollection,
+	matcherLlvmBitcode,
+	matcherParquet,
+	matcherAvro,
+	matcherPem,
+	matcherPe,
+	matcherIso9660,
+	matcherBmp,
+	matcherPdf,
+	matcherTiff,
+	matcherMp3,
+	matcherMp4,
+	matcherRar,
+	matcher7zip,
+	matcherIco,
+	matcherSqlite,
+	matcherPcapng,
+	matcherPcap,
+	matcherFlac,
+	matcherTdf,
+	matcherTdef,
+	matcherCab,
+	matcherPsd,
+	matcherAvi,
+	matcherOle,
+	matcherWebp,
+	matcherRtf,
+	matcherHtml,
+	matcherXml,
+	matcherText,
+}
